Typescript fixes
