+++ v2/pink-sb/src/lib/DirectoryPicker/index.ts
@@ -1,0 +1,8 @@
+export type Directory = {
+    id: string;
+    name: string;
+    fullPath: string;
+    fileCount: number;
+    thumbnailUrl: string;
+    subDirectories: Directory[];
+};
