+++ v2/pink-sb/src/lib/selector/Radio.svelte
@@ -8,6 +8,7 @@
     } & Partial<{
             size: 'small' | 'medium';
             label: string;
+            radioInput: HTMLInputElement;
         }>;
 
     export let group: $$Props['group'];
@@ -15,7 +16,7 @@
     export let size: $$Props['size'] = 'medium';
     export let id: string | undefined = undefined;
     export let label: $$Props['label'] = undefined;
-    export let radioInput;
+    export let radioInput: $$Props['radioInput'];
 </script>
 
 <Base {label} {id}>
